add the 'fitBounds()' functionality
This function takes 2 sets of arguments, the North-East Lat/Lng and the
South-West Lat/Lng, which are used to compute a LatLngBounds() which is then
passed to the gmap fitBounds() method.

The result is a map with the correct zoom to show the NE-SW region.

As a possible improvement, we could automatically store the NE/SW corners for
every new object added to the map (via gmplot methods) and then just
'requesting' to fit those bounds by flipping a True/False flag.